Fix crash on loading conversation
The switchMap was causing a crash because it was trying to use conversation as an observable, even though it is not. (caused by ca50c4d4ba893c77fb1c81ca1099a030a12b5b4a) There is no need for this code, because the whole function is now an observable.